pyproject.toml
[project]
name = "code-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "crewai>=0.120.1",
    "langgraph>=0.4.3",
    "mcp>=1.8.1",
    "mistune>=3.1.3",
    "python-dotenv>=1.1.0",
]
